pyproject.toml
[project]
name = "rag-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.7",
    "gradio>=6.9.0",
    "llama-index>=0.14.16",
    "llama-index-core>=0.14.17",
    "llama-index-embeddings-cohere>=0.8.0",
    "llama-index-llms-cohere>=0.8.0",
    "llama-index-vector-stores-pinecone>=0.8.0",
    "pinecone-client>=6.0.0",
    "pip-system-certs>=5.3",
    "pyreadline3>=3.5.4",
    "python-dotenv>=1.2.2",
]
